refactor: hooks의 queries 하위 폴더 생성, query 훅과 분리
