+++ src/hooks/queries/useDeleteCartItem.ts
@@ -2,7 +2,7 @@ import { deleteCartItem } from '@/api';
 import { ERROR } from '@/constant/message';
 import { MUTATION_KEY, QUERY_KEY } from '@/constant/queryKey';
 import { useMutation, useQueryClient } from '@tanstack/react-query';
-import { useToast } from './useToast';
+import { useToast } from '../useToast';
 
 const useDeleteCartItem = () => {
   const client = useQueryClient();
