+++ src/hooks/test/usePatchCartItemQuantity.test.tsx
@@ -1,7 +1,7 @@
 import { act, renderHook, waitFor } from '@testing-library/react';
 import { queryWrapper } from './wrapper.util';
-import useCartItems from '../useCartItems';
-import usePatchCartItemQuantity from '../usePatchCartItemQuantity';
+import useCartItems from '../queries/useCartItems';
+import usePatchCartItemQuantity from '../queries/usePatchCartItemQuantity';
 
 describe('usePatchCartItemQuantity 훅 테스트', () => {
   it('장바구니 ID 11294와 수량 2로 해당 상품의 장바구니 수량을 변경시킨다', async () => {
